+++ t/009-apicast-caching.t
@@ -160,3 +160,62 @@ apicast cache miss key: 1:one-key:usage%5Bhits%5D=1
 apicast cache write key: 1:one-key:usage%5Bhits%5D=1
 apicast cache miss key: 2:two-id:two-key:usage%5Bhits%5D=2
 apicast cache write key: 2:two-id:two-key:usage%5Bhits%5D=2
+
+=== TEST 7: call to backend is cached
+First call is done synchronously and the second out of band.
+--- http_config
+  include $TEST_NGINX_UPSTREAM_CONFIG;
+  lua_package_path "$TEST_NGINX_LUA_PATH";
+  init_by_lua_block {
+    require('configuration_loader').mock({
+      services = {
+        {
+          id = 42,
+          backend_version = 'oauth',
+          proxy = {
+            credentials_location = "query",
+            api_backend = "http://127.0.0.1:$TEST_NGINX_SERVER_PORT/api-backend/",
+            proxy_rules = {
+              { pattern = '/', http_method = 'GET', metric_system_name = 'hits', delta = 2 }
+            }
+          }
+        }
+      }
+    })
+  }
+  lua_shared_dict api_keys 10m;
+--- config
+  include $TEST_NGINX_APICAST_CONFIG;
+
+  set $backend_endpoint 'http://127.0.0.1:$TEST_NGINX_SERVER_PORT';
+  set $backend_authentication_type 'service_token';
+  set $backend_authentication_value 'token-value';
+
+  location /transactions/oauth_authrep.xml {
+    content_by_lua_block { ngx.exit(200) }
+  }
+
+  location /api-backend/ {
+     echo 'yay, api backend';
+  }
+
+  location ~ /test/(.+) {
+    proxy_pass $scheme://127.0.0.1:$server_port/$1$is_args$args;
+    proxy_set_header Host localhost;
+  }
+
+  location = /t {
+    echo_subrequest GET /test/one -q access_token=value;
+    echo_subrequest GET /test/two -q access_token=value;
+  }
+--- request
+GET /t
+--- response_body
+yay, api backend
+yay, api backend
+--- error_code: 200
+--- grep_error_log eval: qr/apicast cache (?:hit|miss|write) key: [^,\s]+/
+--- grep_error_log_out
+apicast cache miss key: 42:value:usage%5Bhits%5D=2
+apicast cache write key: 42:value:usage%5Bhits%5D=2
+apicast cache hit key: 42:value:usage%5Bhits%5D=2